@@ -3,18 +3,18 @@
   "version": "0.0.0",
   "license": "MIT",
   "scripts": {
-    "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
+    "test-lib": "npm run build && npm pack ./dist && npm install ngx-ui-scroll-0.0.0.tgz && npm run start",
+    "build": "node build.js",
+    "start": "ng serve --delete-output-path=false",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e"
+    "e2e": "ng e2e",
+    "ng": "ng"
   },
   "private": true,
   "dependencies": {
     "@angular/animations": "^5.1.2",
     "@angular/common": "^5.1.2",
-    "@angular/compiler": "^5.1.2",
     "@angular/core": "^5.1.2",
     "@angular/forms": "^5.1.2",
     "@angular/http": "^5.1.2",
@@ -22,17 +22,21 @@
     "@angular/platform-browser-dynamic": "^5.1.2",
     "@angular/router": "^5.1.2",
     "core-js": "^2.5.3",
+    "ngx-ui-scroll": "file:ngx-ui-scroll-0.0.0.tgz",
     "rxjs": "^5.5.6",
     "zone.js": "^0.8.14"
   },
   "devDependencies": {
     "@angular/cli": "^1.6.2",
+    "@angular/compiler": "^5.1.2",
     "@angular/compiler-cli": "^5.1.2",
     "@angular/language-service": "^5.1.2",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "^6.0.95",
+    "chalk": "^2.3.0",
     "codelyzer": "~3.2.0",
+    "cpx": "^1.5.0",
     "jasmine-core": "~2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
     "karma": "~1.7.0",
@@ -42,8 +46,15 @@
     "karma-jasmine": "^1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
     "protractor": "~5.1.2",
+    "renamer": "^0.6.1",
+    "rollup": "^0.53.2",
+    "rollup-plugin-license": "^0.5.0",
+    "rollup-plugin-node-resolve": "^3.0.0",
+    "rollup-plugin-sourcemaps": "^0.4.2",
+    "shelljs": "^0.7.8",
     "ts-node": "~3.2.0",
-    "tslint": "~5.7.0",
-    "typescript": "~2.4.2"
+    "tslint": "^5.7.0",
+    "typescript": "^2.5.3",
+    "uglify-js": "^3.3.4"
   }
 }
