@@ -2,7 +2,10 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 
 import { AppComponent } from './app.component';
-import { UiScrollModule } from '../../src';
+//import { UiScrollModule } from '../../src/ngx-ui-scroll';
+//import { UiScrollModule } from '../../dist/bundles/ngx-ui-scroll.umd.js';
+//import { UiScrollModule } from '../../public_api';
+import { UiScrollModule } from 'ngx-ui-scroll';
 
 @NgModule({
   declarations: [AppComponent],
