@@ -1,10 +1,10 @@
-import { ElementRef } from '@angular/core'
+import {ElementRef} from '@angular/core';
 
 export default class Elements {
 
-  static viewport;
-  static paddingTop;
-  static paddingBottom;
+  static viewport = null;
+  static paddingTop = null;
+  static paddingBottom = null;
 
   static initialize(elementRef: ElementRef) {
     Elements.viewport = elementRef.nativeElement;
